// src/components/Transactions/types.ts
export type TransactionItem = {
  id: string | number;
  name: string;
  category: string;
  value: number;
  date: string;
  type: string;
  senderName: string;
  senderAcc: string;
  senderBank: string;
  reciever: string;
  recieverAcc: string;
  recieverBank: string;
  payType: string;
  description: string;
  docNumber: string;
};